+++ src/worker.rs
@@ -209,7 +209,8 @@ pub async fn worker_main(goose_attack: &GooseAttack) -> GooseAttack {
 
     // Divide hatch_rate amongst all workers.
     let hatch_rate = 1.0
-        / (worker_goose_attack.configuration.hatch_rate as f32 / (config.expect_workers as f32));
+        // Hatch rate is required here, so unwrap() is safe.
+        / (worker_goose_attack.configuration.hatch_rate.unwrap() as f32 / (config.expect_workers as f32));
     info!(
         "[{}] prepared to start 1 user every {:.2} seconds",
         worker_id, hatch_rate,