+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.10.1-dev
+ - rework `hatch_rate` to be stored in an `Option<usize>` as it can be `None` on a Worker
 
 ## 0.10.0 Sep 13, 2020
  - default to resetting statistics, disable with `--no-reset-stats`, display spawning statistics before resetting