+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - remove redundant `GooseAttack.users` instead using the `Option<usize>` in `configuration`
  - improve bounds handling of defaults, generate errors for invalid values
  - properly handle early shutdown of Gaggle distributed load test from Worker process
+ - Manager starts timing Gaggle distributed load test only after all Workers start
 
 ## 0.10.0 Sep 13, 2020
  - default to resetting statistics, disable with `--no-reset-stats`, display spawning statistics before resetting